ci: enable docker image publications on tags (#116)
Enable publishing a Caddy docker image that contains the
cloudflare plugin installed. The image publication part of
CI will only run when a tag is pushed.

The approach followed in the Dockerfile here is to base
the runtime image on a pure Alpine image vs caddy:alpine
because with the latter approach copying the binary in
the image increases the image size instead of keeping it
the same (due to how container image layers work).